feat(F-417): optimize services
