+++ frontend/src/app/[locale]/(main)/history/components/PreviousSessions.tsx
@@ -4,7 +4,7 @@ import { ChevronDown, Download, Trash2 } from 'lucide-react';
 import { Button } from '@/components/ui/Button';
 import { useEffect, useState } from 'react';
 import { useTranslations } from 'next-intl';
-import { clearAllSessions } from '@/services/client/SessionService';
+import { clearAllSessions, sessionService } from '@/services/SessionService';
 import {
   AlertDialog,
   AlertDialogAction,
@@ -18,10 +18,9 @@ import {
 } from '@/components/ui/AlertDialog';
 import { showErrorToast, showSuccessToast } from '@/lib/toast';
 import { SessionPaginated, SessionFromPagination } from '@/interfaces/models/Session';
-import { api } from '@/services/client/Api';
-import { sessionService } from '@/services/server/SessionService';
 import Link from 'next/link';
 import EmptyListComponent from '@/components/common/EmptyListComponent';
+import { api } from '@/services/ApiClient';
 
 export default function PreviousSessions({
   limit,
@@ -65,7 +64,7 @@ export default function PreviousSessions({
   const handleDeleteAll = async () => {
     try {
       setIsDeleting(true);
-      await clearAllSessions();
+      await clearAllSessions(api);
 
       setSessionsStorage([]);
       setTotalSessionsCount(0);
