+++ frontend/src/app/[locale]/(main)/history/page.tsx
@@ -2,9 +2,9 @@ import { Suspense } from 'react';
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/props/MetadataProps';
-import { sessionService } from '@/services/server/SessionService';
-import { api } from '@/services/server/Api';
-import { UserProfileService } from '@/services/server/UserProfileService';
+import { sessionService } from '@/services/SessionService';
+import { UserProfileService } from '@/services/UserProfileService';
+import { api } from '@/services/ApiServer';
 import HistoryHeader from './components/HistoryHeader';
 import HistoryStats from './components/HistoryStats';
 import PreviousSessions from './components/PreviousSessions';
@@ -18,7 +18,7 @@ export async function generateMetadata({ params }: MetadataProps): Promise<Metad
 export default async function HistoryPage() {
   const PAGE_SIZE = 3;
   const sessions = await sessionService.getPaginatedSessions(api, 1, PAGE_SIZE);
-  const userStatsData = await UserProfileService.getUserStats();
+  const userStatsData = await UserProfileService.getUserStats(api);
 
   return (
     <Suspense fallback={<Loading />}>
