+++ frontend/src/app/[locale]/(main)/new-conversation-scenario/page.tsx
@@ -1,7 +1,8 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/props/MetadataProps';
-import { conversationScenarioService } from '@/services/server/ConversationScenarioService';
+import { conversationScenarioService } from '@/services/ConversationScenarioService';
+import { api } from '@/services/ApiServer';
 import ConversationScenarioForm from './components/ConversationScenarioForm';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
@@ -10,6 +11,6 @@ export async function generateMetadata({ params }: MetadataProps): Promise<Metad
 }
 
 export default async function ConversationScenarioPage() {
-  const categories = await conversationScenarioService.getConversationCategories();
+  const categories = await conversationScenarioService.getConversationCategories(api);
   return <ConversationScenarioForm categoriesData={categories.data} />;
 }
