+++ frontend/src/app/[locale]/(main)/preparation/[id]/components/CreateSessionButton.tsx
@@ -1,12 +1,13 @@
 'use client';
 
 import { Button } from '@/components/ui/Button';
-import { sessionService } from '@/services/client/SessionService';
+import { sessionService } from '@/services/SessionService';
 import { Play } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/navigation';
 import { showErrorToast } from '@/lib/toast';
 import { useState } from 'react';
+import { api } from '@/services/ApiClient';
 
 interface CreateSessionButtonProps {
   scenarioId: string;
@@ -23,7 +24,7 @@ export const CreateSessionButton = ({ scenarioId }: CreateSessionButtonProps) =>
     setIsSubmitting(true);
 
     try {
-      const { data } = await sessionService.createSession(scenarioId);
+      const { data } = await sessionService.createSession(api, scenarioId);
       router.push(`/simulation/${data.id}`);
     } catch (error) {
       showErrorToast(error, t('sessionCreationError'));
