+++ frontend/src/app/[locale]/(main)/settings/page.tsx
@@ -1,8 +1,9 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/props/MetadataProps';
-import { UserProfileService } from '@/services/server/UserProfileService';
+import { UserProfileService } from '@/services/UserProfileService';
 import { Suspense } from 'react';
+import { api } from '@/services/ApiServer';
 import Settings from './components/Settings';
 import SettingsLoadingPage from './loading';
 
@@ -12,7 +13,7 @@ export async function generateMetadata({ params }: MetadataProps): Promise<Metad
 }
 
 export default function SettingsPage() {
-  const userProfile = UserProfileService.getUserProfile();
+  const userProfile = UserProfileService.getUserProfile(api);
   return (
     <Suspense fallback={<SettingsLoadingPage />}>
       <Settings userProfile={userProfile}></Settings>
