+++ frontend/src/services/ConversationScenarioService.ts
@@ -2,9 +2,19 @@ import {
   ConversationScenarioResponse,
   ConversationScenario,
 } from '@/interfaces/models/ConversationScenario';
-import { api } from './Api';
+import { AxiosInstance } from 'axios';
 
-const createConversationScenario = async (scenario: ConversationScenario) => {
+const getConversationCategories = async (api: AxiosInstance) => {
+  try {
+    const response = await api.get('/conversation-categories');
+    return response;
+  } catch (error) {
+    console.error('Error fetching conversation categories:', error);
+    throw error;
+  }
+};
+
+const createConversationScenario = async (api: AxiosInstance, scenario: ConversationScenario) => {
   try {
     const response = await api.post<ConversationScenarioResponse>('/conversation-scenario', {
       ...scenario,
@@ -16,7 +26,7 @@ const createConversationScenario = async (scenario: ConversationScenario) => {
   }
 };
 
-const getPreparation = async (id: string) => {
+const getPreparation = async (api: AxiosInstance, id: string) => {
   try {
     const response = await api.get(`/conversation-scenario/${id}/preparation`);
     return response;
@@ -27,6 +37,7 @@ const getPreparation = async (id: string) => {
 };
 
 export const conversationScenarioService = {
+  getConversationCategories,
   getPreparation,
   createConversationScenario,
 };
