+++ frontend/src/services/VerificationService.ts
@@ -1,12 +1,12 @@
 import { SendVerificationCodeRequest, VerifyCodeRequest } from '@/interfaces/models/Auth';
-import { api } from './Api';
+import { AxiosInstance } from 'axios';
 
-const sendVerificationCode = async (data: SendVerificationCodeRequest) => {
+const sendVerificationCode = async (api: AxiosInstance, data: SendVerificationCodeRequest) => {
   const response = await api.post('/auth/send-verification', data);
   return response.data;
 };
 
-const verifyCode = async (data: VerifyCodeRequest) => {
+const verifyCode = async (api: AxiosInstance, data: VerifyCodeRequest) => {
   const response = await api.post('/auth/verify-code', data);
   return response.data;
 };
